[refa] Move isShowN2000HeightSystem from depthStyles.ts to features.ts
