+++ src/components/features.ts
@@ -281,3 +281,13 @@ export type ForecastFeatureProperties = {
   pilotPlaceId?: number;
   forecastItems: ForecastItem[];
 };
+
+export function isShowN2000HeightSystem(props: AreaFeatureProperties | LineFeatureProperties): boolean {
+  if (props.n2000HeightSystem !== undefined) {
+    return props.n2000HeightSystem;
+  } else if (props.isN2000 !== undefined) {
+    return props.isN2000;
+  } else {
+    return (props.referenceLevel && props.referenceLevel.indexOf('N2000') !== -1) || !!props.n2000depth || !!props.n2000draft;
+  }
+}
