+++ src/components/popup/AreaPopupContent.tsx
@@ -4,9 +4,8 @@ import { useTranslation } from 'react-i18next';
 import './popup.css';
 import { Link } from 'react-router-dom';
 import { Lang } from '../../utils/constants';
-import { AreaFeatureProperties } from '../features';
+import { AreaFeatureProperties, isShowN2000HeightSystem } from '../features';
 import InfoIcon from '../../theme/img/info.svg?react';
-import { isShowN2000HeightSystem } from '../layerStyles/depthStyles';
 import { PopupProperties } from '../mapOverlays/MapOverlays';
 import { clearClickSelectionFeatures } from './selectInteraction';
 import CloseButton from './CloseButton';
