+++ src/components/popup/LinePopupContent.tsx
@@ -4,9 +4,8 @@ import { useTranslation } from 'react-i18next';
 import './popup.css';
 import { Link } from 'react-router-dom';
 import { Lang } from '../../utils/constants';
-import { LineFeatureProperties } from '../features';
+import { LineFeatureProperties, isShowN2000HeightSystem } from '../features';
 import InfoIcon from '../../theme/img/info.svg?react';
-import { isShowN2000HeightSystem } from '../layerStyles/depthStyles';
 import { PopupProperties } from '../mapOverlays/MapOverlays';
 import dvkMap from '../DvkMap';
 import { clearClickSelectionFeatures } from './selectInteraction';
